+++ functions/public/Add-PSWorkItemCategory.ps1
@@ -1,5 +1,5 @@
 Function Add-PSWorkItemCategory {
-    [cmdletbinding(SupportsShouldProcess,DefaultParameterSetName = "file")]
+    [cmdletbinding(SupportsShouldProcess)]
     Param(
         [Parameter(
             Position = 0,
@@ -17,26 +17,21 @@ Function Add-PSWorkItemCategory {
             HelpMessage = "Specify a category comment or description",
             ValueFromPipelineByPropertyName
         )]
-        [alias("description")]
-        [string]$Comment,
+        [string]$Description,
 
-        [Parameter(HelpMessage = "The path to the PSWorkitem SQLite database file. It should end in .db",ParameterSetName="file")]
+        [Parameter(HelpMessage = "The path to the PSWorkitem SQLite database file. It should end in .db")]
         [ValidateNotNullOrEmpty()]
         [ValidatePattern("\.db$")]
         [ValidateScript({
-                $parent = Split-Path -Path $_ -Parent
-                if (Test-Path $parent) {
+                if (Test-Path $_) {
                     Return $True
                 }
                 else {
-                    Throw "Failed to validate the parent path $parent."
+                    Throw "Failed to validate $_"
                     Return $False
                 }
             })]
         [string]$Path = $PSWorkItemPath,
-
-        [Parameter(HelpMessage = "Specify an existing database connection",ParameterSetName="connection")]
-        [System.Data.SQLite.SQLiteConnection]$Connection,
 
         [Parameter(HelpMessage = "Force overwriting an existing category")]
         [switch]$Force,
@@ -48,30 +43,23 @@ Function Add-PSWorkItemCategory {
         Write-Debug "Using bound parameters"
         $PSBoundParameters | Out-String | Write-Debug
 
-        if ($pscmdlet.ParameterSetName -eq 'file') {
-
-            Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] $($myinvocation.mycommand): Opening a connection to $Path"
-            Try {
-                $conn = Open-MySQLiteDB -Path $Path -ErrorAction Stop
-                $conn | Out-String | Write-Debug
+        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] $($myinvocation.mycommand): Opening a connection to $Path"
+        Try {
+            $conn = Open-MySQLiteDB -Path $Path -ErrorAction Stop
+            $conn | Out-String | Write-Debug
 
         }
         Catch {
             Throw "$($myinvocation.mycommand): Failed to open the database $Path"
         }
-    }
-    else {
-        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] $($myinvocation.mycommand): Using an existing connection"
-        $conn = $Connection
-    }
 
-     #parameters to splat to Invoke-MySQLiteQuery
-     $splat = @{
-        Connection  = $conn
-        KeepAlive   = $true
-        Query       = ""
-        ErrorAction = "Stop"
-    }
+        #parameters to splat to Invoke-MySQLiteQuery
+        $splat = @{
+            Connection  = $conn
+            KeepAlive   = $true
+            Query       = ""
+            ErrorAction = "Stop"
+        }
     } #begin
 
     Process {
@@ -87,7 +75,7 @@ Function Add-PSWorkItemCategory {
                 elseif ($test.category -eq $item -AND $Force) {
                     Write-Verbose "$($myinvocation.mycommand): The category $Category already exists and will be overwritten"
                     $splat.Query = "DELETE FROM categories WHERE category = '$item'"
-                    if ($Pscmdlet.ShouldProcess($item,"Remove category")) {
+                    if ($Pscmdlet.ShouldProcess($item, "Remove category")) {
                         Invoke-MySQLiteQuery @splat
                         $ok = $true
                     }
@@ -99,13 +87,13 @@ Function Add-PSWorkItemCategory {
                 Write-Debug "$($myinvocation.mycommand): Connection state is $($conn.state)"
                 if ($ok) {
                     Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($myinvocation.mycommand): Adding category $Item"
-                    $splat.query = "INSERT INTO categories (category,comment) VALUES ('$item','$comment')"
+                    $splat.query = "INSERT INTO categories (category,description) VALUES ('$item','$Description')"
                     If ($pscmdlet.ShouldProcess($item)) {
                         Invoke-MySQLiteQuery @splat
                         if ($Passthru) {
                             $splat.query = "Select * from categories where category = '$item'"
                             Invoke-MySQLiteQuery @splat | ForEach-Object {
-                                [PSWorkItemCategory]::New($_.category, $_.comment)
+                                [PSWorkItemCategory]::New($_.category, $_.Description)
                             }
                         } #passthru
                     } #Whatif
@@ -120,5 +108,4 @@ Function Add-PSWorkItemCategory {
         Close-MySQLiteDB -Connection $conn
         Write-Verbose "[$((Get-Date).TimeofDay) END    ] $($myinvocation.mycommand): Ending"
     } #end
-
 }