+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog for PSWorkItem
 
+## 0.2.0
+
++ Added `Add-PSWorkitemCategory`,`Get-PSWorkItemCategory`, and `Remove-PSWorkItemCategory`.
++ Added `Get-PSWorkingItemDatabase`.
++ Added class definition for `PSWorkItemDatabase`.
++ Added format file `psworkitemdatabase.format.ps1xml`.
++ Updated `Initialize-PSWorkItemDatabase` to add default categories.
++ Improved verbose messaging.
+
 ## 0.1.0
 
 + Initial files